pyproject.toml
[project]
name = "playwright-py-sample"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
  "playwright>=1.50.0",
  "pytest>=8.3.5",
  "pytest-playwright>=0.7.0",
  "pytest-xdist>=3.6.1",
  "python-dotenv>=1.0.1",
]

[tool.pytest.ini_options]
addopts = "-v -s"
usefixtures = "setup_env"
